[tests] Move concurrency tests to new file
